Move resources arguments to input.args
