+++ lib/core/network/removeErrorStack.js
@@ -46,7 +46,9 @@ module.exports = data => {
       data.content.error.stack = undefined;
     }
     else {
-      data.content.error.stack = data.content.error.stack.split('\n').map(hilightUserCode).join('\n');
+      data.content.error.stack = data.content.error.stack
+        ? data.content.error.stack.split('\n').map(hilightUserCode).join('\n')
+        : undefined;
     }
   }
